+++ crates/core/src/errors.rs
@@ -56,6 +56,9 @@ pub enum ProtoError {
     #[error("TOML failure for {0}: {1}")]
     Toml(PathBuf, String),
 
+    #[error("Unable to unpack {0}, unsupported archive format {1}.")]
+    UnsupportedArchiveFormat(PathBuf, String),
+
     #[error("Unable to install {0}, unsupported architecture {1}.")]
     UnsupportedArchitecture(String, String),