+++ crates/cli/src/bin.rs
@@ -23,6 +23,9 @@ async fn main() {
             pin,
             passthrough,
         } => commands::install(tool, semver, pin, passthrough).await,
+        Commands::InstallGlobal { tool, dependency } => {
+            commands::install_global(tool, dependency).await
+        }
         Commands::Global { tool, semver } => commands::global(tool, semver).await,
         Commands::List { tool } => commands::list(tool).await,
         Commands::ListRemote { tool } => commands::list_remote(tool).await,